refactored helper functions
