+++ waterloop-site/src/pages/TeamPage/TeamPage.tsx
@@ -2,11 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 
 import getKeyByValue from '../../utils/getKeyByValue'
-import {
-  sortProfiles,
-  generateMembersQuery,
-  generateFiltersQuery
-} from './utils'
+import { generateMembersQuery, generateFiltersQuery } from './utils'
 
 import { ProfileSection } from '../../components/Profiles'
 import TeamProfileFilter from '../../components/TeamFilter'
@@ -24,20 +20,73 @@ const Page = styled.div`
   }
 `
 
+// Check if member has crucial missing fields
+const isProfileComplete = (member: any) => {
+  return !(!member.memberType || member.memberType.length === 0 || !member.subteams || member.subteams.length === 0)
+}
+
+// Identify any new teams and register them into the Map/Array
+const identifyNewTeams = (member: any, map: any, arr: Array<Array<Object>>) => {
+  // Team Leads will all be in in the first subarray regardless of their subteam
+  // Do not make a new subarray for them since it might remain empty
+  if (member.memberType.name === "Technical Director") return
+
+  // Create a new subarray and register index in map for each new subteam
+  member.subteams.forEach((team: string) => {
+    if (!map.has(team)) {
+      map.set(team, arr.length)
+      arr.push([])
+    }
+  })
+}
+
+// Insert members into the appropriate subarray(s)
+const insertMember = (member: any, map: any, arr: Array<Array<Object>>) => {
+  // Insert all Team Leads into the first subarray
+  if (member.memberType.name === "Technical Director") {
+    arr[0].push(member)
+  }
+  else {
+    // Insert Subteam leads/members into corresonding subarray(s)
+    member.subteams.forEach((team: string) => {
+      // Insert Subteam Lead members to front of subarray
+      if (member.memberType.name === "Subteam Lead") {
+        arr[map.get(team)].unshift(member)
+      } else {
+        arr[map.get(team)].push(member)
+      }
+    })
+  }
+}
+
+// Format 1d array of members into 2d array: [[team leads], [subteam 1 members], ... ]
+const sortProfiles = (members: any) => {
+  let sortedTeams = [[]] as Array<Array<any>>
+  let subteamIndex = new Map()
+
+  // Insert members into their corresponding subarrays within sortedTeams
+  members.forEach((member: any, key: number) => {
+    if (isProfileComplete (member)) {
+      identifyNewTeams(member, subteamIndex, sortedTeams)
+      insertMember(member, subteamIndex, sortedTeams)
+    }
+  })
+
+  return [sortedTeams, subteamIndex]
+}
+
 export default class TeamPage extends React.Component<any, any> {
   constructor(props: any) {
     super(props)
     this.state = {
       teamFilters: Array(5).fill(false),
-      memberData: [[]] as any,
       toggleOpen: false,
-      subteamMap: {} as any,
-      subteamServerId: new Map() as Map<string, string>
+      memberData: [[]] as any,
+      subteamIdMap: new Map() as Map<string, string>,
+      subteamMap: {} as any
     }
 
     this.fetchSubteams();
-
-    // TODO: Replace the arg with a prop once the intro team page is done.
     this.fetchProfiles();
   }
 
@@ -49,9 +98,9 @@ export default class TeamPage extends React.Component<any, any> {
       .then(res => {
         console.log("finished fetching subteam data", res.body.subteams)
         res.body.subteams.forEach((team: any) => {
-          let newMap = this.state.subteamServerId
+          let newMap = this.state.subteamIdMap
           newMap.set(team._id, team.name)
-          this.setState({subteamServerId: newMap})
+          this.setState({subteamIdMap: newMap})
         })
       })
       .catch(err => {
@@ -119,7 +168,7 @@ export default class TeamPage extends React.Component<any, any> {
         />}
 
         {subteams.length > 0 && subteams.map((subteam: any, i: number) => {
-          const teamName = `${this.state.subteamServerId.get(getKeyByValue(this.state.subteamMap, i+1))}`
+          const teamName = `${this.state.subteamIdMap.get(getKeyByValue(this.state.subteamMap, i+1))}`
           return <ProfileSection
             key={i}
             title={teamName}
